requirements.txt
python-dateutil>=2.9.0
openai>=1.46.0
anthropic>=0.34.2
python-telegram-bot>=21.4
discord.py>=2.4.0
fastapi>=0.115.0
uvicorn>=0.30.6
